deleted component call-list copied code to pages call-list
